+++ app/src/main/java/org/radarcns/android/DeviceState.java
@@ -2,6 +2,7 @@ package org.radarcns.android;
 
 import android.os.Parcelable;
 
+/** Current state of a wearable device. */
 public interface DeviceState extends Parcelable {
     DeviceStatusListener.Status getStatus();
 }
