+++ app/src/main/java/org/radarcns/data/AvroDecoder.java
@@ -4,9 +4,9 @@ import org.apache.avro.Schema;
 
 import java.io.IOException;
 
-/** Encode Avro values with a given encoder */
+/** Decode Avro values with a given encoder */
 public interface AvroDecoder {
-    /** Create a new writer. This method is thread-safe */
+    /** Create a new reader. This method is thread-safe, but the class it returns is not. */
     <T> AvroReader<T> reader(Schema schema, Class<T> clazz) throws IOException;
 
     interface AvroReader<T> {
