+++ app/src/main/java/org/radarcns/data/AvroEncoder.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 
 /** Encode Avro values with a given encoder */
 public interface AvroEncoder {
-    /** Create a new writer. This method is thread-safe */
+    /** Create a new writer. This method is thread-safe, but the class it returns is not. */
     <T> AvroWriter<T> writer(Schema schema, Class<T> clazz) throws IOException;
 
     interface AvroWriter<T> {
