+++ app/src/main/java/org/radarcns/data/Record.java
@@ -1,5 +1,13 @@
 package org.radarcns.data;
 
+/**
+ * A single data record.
+ *
+ * The time it gets created is stored as {@link #milliTimeAdded}.
+ *
+ * @param <K> key type
+ * @param <V> value type
+ */
 public class Record<K, V> {
     public final long offset;
     public final K key;
